refactor: thin admin pickup page into subcomponents (1.0.9)
app/admin/pickup/page.tsx drops from 814 to 344 lines, keeping state,
data fetching, handlers, and derived values as a container.

Extracted:
- constants.ts: types, label maps, and pure helpers (formatDate,
  getDaysInMonth, getTimeDisplay)
- components/: UpcomingTab, DatesTab, ReservationsTab, ManualAddModal
  as presentational components driven by props

UI and behavior unchanged; no API change. tsc --noEmit, next build,
and 25 webhook tests all pass.
