+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "antnest-chatbot",
-  "version": "1.0.8",
+  "version": "1.0.9",
   "description": "",
   "main": "index.js",
   "scripts": {
